moved some parts of the code to avoid cycles
