+++ pkg/internal/export/expire/expiry_map.go
@@ -1,4 +1,4 @@
-package export
+package expire
 
 import (
 	"strings"
